refactor.
The path to the ebs directory is given as a command line argument to avoid
environment variables.

Don't buffer tasks unless we really need to (for list and guess).

Parse integers properly with strtoimax instead of scanf so we can check for
overflow in add. Reading from task.tsv still uses scanf.